Move Component to UML.classes
Remove Node-Component grouping.
