+++ gaphor/UML/classes/classestoolbox.py
@@ -83,6 +83,18 @@ classes = ToolSection(
             handle_index=SE,
         ),
         ToolDef(
+            "toolbox-component",
+            gettext("Component"),
+            "gaphor-component-symbolic",
+            "o",
+            new_item_factory(
+                diagramitems.ComponentItem,
+                UML.Component,
+                config_func=namespace_config,
+            ),
+            handle_index=SE,
+        ),
+        ToolDef(
             "toolbox-containment",
             gettext("Containment"),
             "gaphor-containment-symbolic",
