+++ gaphor/UML/classes/component.py
