+++ gaphor/UML/classes/propertypages.glade
@@ -902,4 +902,33 @@ Use &lt;b&gt;-&lt;/b&gt;/&lt;b&gt;=&lt;/b&gt; to move items up or down.</propert
       </object>
     </child>
   </object>
+  <object class="GtkBox" id="component-editor">
+    <property name="visible">1</property>
+    <property name="margin-top">6</property>
+    <property name="orientation">vertical</property>
+    <property name="spacing">6</property>
+    <child>
+      <object class="GtkBox">
+        <property name="visible">1</property>
+        <child>
+          <object class="GtkLabel">
+            <property name="visible">1</property>
+            <property name="label" translatable="yes">Indirectly Instantiated</property>
+          </object>
+        </child>
+        <child>
+          <object class="GtkSwitch" id="indirectly-instantiated">
+            <property name="visible">1</property>
+            <property name="can-focus">1</property>
+            <signal name="notify::active" handler="indirectly-instantiated-changed" swapped="no"/>
+          </object>
+          <packing>
+            <property name="fill">0</property>
+            <property name="pack-type">end</property>
+            <property name="position">1</property>
+          </packing>
+        </child>
+      </object>
+    </child>
+  </object>
 </interface>
