+++ gaphor/UML/deployments/__init__.py
@@ -5,8 +5,7 @@ from gaphor.UML.deployments import (
     deploymentsgrouping,
 )
 from gaphor.UML.deployments.artifact import ArtifactItem
-from gaphor.UML.deployments.component import ComponentItem
 from gaphor.UML.deployments.connector import ConnectorItem
 from gaphor.UML.deployments.node import NodeItem
 
-__all__ = ["ArtifactItem", "ComponentItem", "ConnectorItem", "NodeItem"]
+__all__ = ["ArtifactItem", "ConnectorItem", "NodeItem"]
