+++ gaphor/UML/deployments/propertypages.glade
@@ -14,35 +14,6 @@ Edit, from the popup menu, will allow you to add cell
 renderers and such.</property>
     <property name="xalign">0</property>
   </object>
-  <object class="GtkBox" id="component-editor">
-    <property name="visible">1</property>
-    <property name="margin-top">6</property>
-    <property name="orientation">vertical</property>
-    <property name="spacing">6</property>
-    <child>
-      <object class="GtkBox">
-        <property name="visible">1</property>
-        <child>
-          <object class="GtkLabel">
-            <property name="visible">1</property>
-            <property name="label" translatable="yes">Indirectly Instantiated</property>
-          </object>
-        </child>
-        <child>
-          <object class="GtkSwitch" id="indirectly-instantiated">
-            <property name="visible">1</property>
-            <property name="can-focus">1</property>
-            <signal name="notify::active" handler="indirectly-instantiated-changed" swapped="no"/>
-          </object>
-          <packing>
-            <property name="fill">0</property>
-            <property name="pack-type">end</property>
-            <property name="position">1</property>
-          </packing>
-        </child>
-      </object>
-    </child>
-  </object>
   <object class="GtkBox" id="information-flow-editor">
     <property name="visible">1</property>
     <property name="margin-top">6</property>
