+++ gaphor/UML/deployments/propertypages.ui
@@ -12,26 +12,6 @@ Edit, from the popup menu, will allow you to add cell
 renderers and such.</property>
     <property name="xalign">0</property>
   </object>
-  <object class="GtkBox" id="component-editor">
-    <property name="margin-top">6</property>
-    <property name="orientation">vertical</property>
-    <property name="spacing">6</property>
-    <child>
-      <object class="GtkBox">
-        <child>
-          <object class="GtkLabel">
-            <property name="label" translatable="yes">Indirectly Instantiated</property>
-          </object>
-        </child>
-        <child>
-          <object class="GtkSwitch" id="indirectly-instantiated">
-            <property name="halign">center</property>
-            <signal name="notify::active" handler="indirectly-instantiated-changed" swapped="no"/>
-          </object>
-        </child>
-      </object>
-    </child>
-  </object>
   <object class="GtkBox" id="information-flow-editor">
     <property name="margin-top">6</property>
     <property name="orientation">vertical</property>
