+++ gaphor/diagram/tests/test_copypaste_grouping.py
@@ -4,28 +4,28 @@ from gaphor import UML
 from gaphor.diagram.copypaste import copy, paste_link
 from gaphor.diagram.grouping import Group
 from gaphor.diagram.tests.fixtures import copy_clear_and_paste_link
-from gaphor.UML.deployments import ComponentItem, NodeItem
+from gaphor.UML.deployments import ArtifactItem, NodeItem
 
 
 @pytest.fixture
-def node_with_component(diagram, element_factory):
+def node_with_artifact(diagram, element_factory):
     node = element_factory.create(UML.Node)
-    comp = element_factory.create(UML.Component)
+    artifact = element_factory.create(UML.Artifact)
     node_item = diagram.create(NodeItem, subject=node)
-    comp_item = diagram.create(ComponentItem, subject=comp)
+    artifact_item = diagram.create(ArtifactItem, subject=artifact)
 
-    Group(node_item, comp_item).group()
-    comp_item.change_parent(node_item)
+    Group(node_item, artifact_item).group()
+    artifact_item.change_parent(node_item)
 
-    assert comp_item.parent is node_item
+    assert artifact_item.parent is node_item
 
-    return node_item, comp_item
+    return node_item, artifact_item
 
 
-def test_copy_paste_of_nested_item(diagram, element_factory, node_with_component):
-    node_item, comp_item = node_with_component
+def test_copy_paste_of_nested_item(diagram, element_factory, node_with_artifact):
+    node_item, artifact_item = node_with_artifact
 
-    buffer = copy({comp_item})
+    buffer = copy({artifact_item})
 
     (new_comp_item,) = paste_link(buffer, diagram, element_factory.lookup)
 
@@ -33,29 +33,29 @@ def test_copy_paste_of_nested_item(diagram, element_factory, node_with_component
 
 
 def test_copy_paste_of_item_with_nested_item(
-    diagram, element_factory, node_with_component
+    diagram, element_factory, node_with_artifact
 ):
-    node_item, comp_item = node_with_component
+    node_item, artifact_item = node_with_artifact
 
-    buffer = copy(set(node_with_component))
+    buffer = copy(set(node_with_artifact))
 
     new_items = paste_link(buffer, diagram, element_factory.lookup)
 
     new_node_item = next(i for i in new_items if isinstance(i, NodeItem))
-    new_comp_item = next(i for i in new_items if isinstance(i, ComponentItem))
+    new_comp_item = next(i for i in new_items if isinstance(i, ArtifactItem))
 
     assert new_comp_item.parent is new_node_item
 
 
 def test_copy_remove_paste_of_item_with_nested_item(
-    diagram, element_factory, node_with_component
+    diagram, element_factory, node_with_artifact
 ):
 
     new_items = copy_clear_and_paste_link(
-        set(node_with_component), diagram, element_factory
+        set(node_with_artifact), diagram, element_factory
     )
 
     new_node_item = next(i for i in new_items if isinstance(i, NodeItem))
-    new_comp_item = next(i for i in new_items if isinstance(i, ComponentItem))
+    new_comp_item = next(i for i in new_items if isinstance(i, ArtifactItem))
 
     assert new_comp_item.parent is new_node_item
