+++ gaphor/diagram/tools/tests/test_dropzone.py
@@ -2,21 +2,21 @@ from gaphas.item import NW, SE
 
 from gaphor import UML
 from gaphor.diagram.tools.dropzone import drop_zone_tool, grow_parent, on_motion
-from gaphor.UML.deployments import ComponentItem, NodeItem
+from gaphor.UML.deployments import ArtifactItem, NodeItem
 
 
 def test_hover_over_drop_zone(diagram, element_factory, view):
     node_item = diagram.create(NodeItem, subject=element_factory.create(UML.Node))
     node_item.width = node_item.height = 200
     comp_item = diagram.create(
-        ComponentItem, subject=element_factory.create(UML.Component)
+        ArtifactItem, subject=element_factory.create(UML.Artifact)
     )
 
     view.request_update((node_item, comp_item))
 
-    tool = drop_zone_tool(view, ComponentItem)
+    tool = drop_zone_tool(view, ArtifactItem)
     view.add_controller(tool)
-    on_motion(tool, 100, 100, ComponentItem)
+    on_motion(tool, 100, 100, ArtifactItem)
 
     selection = view.selection
     assert selection.dropzone_item is node_item
@@ -26,7 +26,7 @@ def test_grow_parent(diagram, element_factory):
     node_item = diagram.create(NodeItem, subject=element_factory.create(UML.Node))
     node_item.width = node_item.height = 200
     comp_item = diagram.create(
-        ComponentItem, subject=element_factory.create(UML.Component)
+        ArtifactItem, subject=element_factory.create(UML.Artifact)
     )
     comp_item.matrix.translate(200, 200)
